+++ packages/animate/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inkset/animate
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @inkset/core@0.1.2
+
 ## 0.1.2
 
 ### Patch Changes